fix(memberships): Fix bug where archived membership option was showing
