+++ ams/memberships/tests/test_forms.py
@@ -1,71 +1,13 @@
-"""Tests for zero-cost membership automatic approval."""
-
-from http import HTTPStatus
-
 import pytest
-from django.urls import reverse
 from django.utils import timezone
 
 from ams.memberships.forms import CreateIndividualMembershipForm
-from ams.memberships.models import IndividualMembership
 from ams.memberships.models import MembershipOptionType
 from ams.users.tests.factories import UserFactory
 
 from .factories import MembershipOptionFactory
 
 pytestmark = pytest.mark.django_db
-
-
-def test_zero_cost_membership_auto_approved(client):
-    """Test that zero-cost memberships are automatically approved."""
-    user = UserFactory()
-    option = MembershipOptionFactory(
-        type=MembershipOptionType.INDIVIDUAL,
-        cost=0,
-    )
-    client.force_login(user)
-
-    resp = client.post(
-        reverse("memberships:apply"),
-        data={
-            "membership_option": option.id,
-            "start_date": timezone.localdate(),
-        },
-        follow=True,
-    )
-
-    assert resp.status_code == HTTPStatus.OK
-    membership = IndividualMembership.objects.get(user=user, membership_option=option)
-    assert membership.start_date == timezone.localdate()
-    assert membership.approved_datetime is not None
-    assert membership.invoice is None
-    assert membership.status().name == "ACTIVE"
-
-
-def test_paid_membership_not_auto_approved(client):
-    """Test that paid memberships are not automatically approved."""
-    user = UserFactory()
-    option = MembershipOptionFactory(
-        type=MembershipOptionType.INDIVIDUAL,
-        cost=99.99,
-    )
-    client.force_login(user)
-
-    resp = client.post(
-        reverse("memberships:apply"),
-        data={
-            "membership_option": option.id,
-            "start_date": timezone.localdate(),
-        },
-        follow=True,
-    )
-
-    assert resp.status_code == HTTPStatus.OK
-    membership = IndividualMembership.objects.get(user=user, membership_option=option)
-    assert membership.start_date == timezone.localdate()
-    assert membership.approved_datetime is None
-    # Note: No invoice in test environment since billing service is not configured
-    assert membership.status().name == "PENDING"
 
 
 def test_zero_cost_membership_form_save_directly():
@@ -140,3 +82,43 @@ def test_zero_cost_membership_shows_active_immediately():
         approved_datetime__isnull=False,
         cancelled_datetime__isnull=True,
     ).exists()
+
+
+def test_archived_membership_option_not_available():
+    """Test that archived membership options cannot be selected."""
+    user = UserFactory()
+    option = MembershipOptionFactory(
+        type=MembershipOptionType.INDIVIDUAL,
+        cost=0,
+        archived=True,
+    )
+
+    form_data = {
+        "membership_option": option.id,
+        "start_date": timezone.localdate(),
+    }
+    form = CreateIndividualMembershipForm(data=form_data, user=user)
+
+    assert not form.is_valid()
+    assert "membership_option" in form.errors
+
+
+def test_archived_membership_option_not_in_queryset():
+    """Test that archived membership options are excluded from form queryset."""
+    user = UserFactory()
+    active_option = MembershipOptionFactory(
+        type=MembershipOptionType.INDIVIDUAL,
+        cost=0,
+        archived=False,
+    )
+    archived_option = MembershipOptionFactory(
+        type=MembershipOptionType.INDIVIDUAL,
+        cost=0,
+        archived=True,
+    )
+
+    form = CreateIndividualMembershipForm(user=user)
+
+    option_ids = [opt.id for opt in form.fields["membership_option"].queryset]
+    assert active_option.id in option_ids
+    assert archived_option.id not in option_ids
